Télécharger et recadrer les données LiDAR pour une zone donnée

Cette fonction interroge l'API STAC de RNCan pour télécharger les données LiDAR (MNE ou MNT) correspondant à une géométrie fournie, puis les recadre à cette zone d’intérêt.

Usage

lidar(polygone, dossier = NULL, mne = FALSE, recent = FALSE)

Arguments

polygone

Un objet `sf` représentant la zone d’intérêt ou un chemin vers un fichier vectoriel (`.shp`, `.gpkg`, etc.)

dossier

Dossier de sortie pour enregistrer les fichiers raster (optionnel)

mne

Logique. Si TRUE, télécharge le MNE (modèle de surface). Sinon, le MNT.

recent

Logique. Si TRUE, conserve uniquement la version la plus récente

Value

Une liste de `SpatRaster` (un par année)