Télécharger et calculer un indice de végétation Sentinel-2 pour une parcelle

Cette fonction interroge le Planetary Computer pour obtenir des images Sentinel-2 L2A, filtre selon la couverture nuageuse, calcule un indice de végétation (NDVI, GNDVI, EVI, KNDVI, NDRE) et retourne une stack raster avec les valeurs de l'indice sélectionné.

Usage

sentinel2(
  polygone,
  dossier = NULL,
  annee_debut = 2018,
  annee_fin = NULL,
  max_nuage = 50,
  mois = 3:11,
  nuage_dans_la_parcelle = 5,
  indice = "NDVI"
)

Arguments

polygone

Un objet `sf` ou un chemin vers un fichier vectoriel (shapefile, geojson, etc.)

dossier

(optionnel) Dossier dans lequel enregistrer les fichiers raster (.tif)

annee_debut

Année de début (par défaut `2018`)

annee_fin

Année de fin (par défaut : année courante)

max_nuage

Pourcentage maximal de nuages (0 à 100) selon les métadonnées

mois

Mois d'intérêt (entiers entre 1 et 12)

nuage_dans_la_parcelle

Proportion maximale de nuages dans la parcelle (0–100), ou `FALSE` pour désactiver

indice

Indice de végétation à calculer (`"NDVI"`, `"GNDVI"`, `"EVI"`, `"KNDVI"`, , `"NDRE"`)

Value

Un objet `SpatRaster` contenant l'indice de végétation par date