Générer une grille orientée selon un polygone

Cette fonction crée une grille de polygones alignée à l'enveloppe orientée minimale d’un polygone. Elle permet d’indiquer un angle, une largeur, une longueur, un coin d’origine, un décalage en quinconce, et de restreindre les cellules à celles qui croisent ou sont contenues dans le polygone.

Usage

st_grille(
  polygone,
  angle = NULL,
  largeur = 100,
  longueur = 100,
  coin = "SO",
  quinconce = FALSE,
  decouper = TRUE,
  points = FALSE
)

Arguments

polygone

Objet `sf` de type POLYGON ou MULTIPOLYGON

angle

Angle d’orientation en degrés. Si NULL, il est déduit automatiquement.

largeur

Largeur des cellules (axe court)

longueur

Longueur des cellules (axe long, optionnel)

coin

Coin de départ : "SO", "SE", "NE" ou "NO"

quinconce

Logique : grille en quinconce (TRUE/FALSE)

decouper

Logique : découper les cellules au bord du polygone

points

Logique : retourner les centroïdes au lieu des polygones

Value

Un objet `sf` avec les colonnes `ligne`, `colonne`, et `geometry`